extend logout base page
